update makefile and format
