+++ api/Makefile
@@ -3,18 +3,36 @@
 
 
 
-.PHONY: api gen watch openapi
+.PHONY: watch migrate
+
+fmt:
+	@go fmt ./...
+
+test:
+	@go test ./...
 
 api:
 	@go build -o ./bin/api ./cmd/api
 
-.PHONY: gen
 gen:
 	@buf generate
 
-.PHONY: watch
 watch:
 	@air --build.cmd "make api" --build.bin "./bin/api"
 
-# openapi:
-# 	@curl http://localhost:8080/openapi.yaml >
+# Applies the specified schema file to the database, similar to Terraform.
+migrate:
+	@atlas schema apply --url "postgres://admin:admin@localhost:5433/happened_db?sslmode=disable" --to "file://sql/schema.sql" --dev-url "docker://postgres/15"
+
+# Connects to the local Postgres database
+connect:
+	@psql "postgres://admin:admin@localhost:5433/happened_db"
+
+#
+db:
+	@docker compose up -d
+
+# Introspects database and returns current schema
+inspect:
+	@atlas schema inspect -u "postgres://admin:admin@localhost:5433/happened_db?sslmode=disable" --format '{{ sql . }}' > schema.sql
+
