--- api/connect.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-psql "postgres://admin:admin@localhost:5433/happened_db"
