--- api/inspect.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-# Introspects database and returns current schema
-atlas schema inspect -u "postgres://admin:admin@localhost:5433/happened_db?sslmode=disable" --format '{{ sql . }}' > schema.sql
-
-
